Separate data flow layer from enhancement layer.
