+++ semanticcpg/src/main/scala/io/shiftleft/layers/EnhancedBaseCreator.scala
@@ -1,25 +1,20 @@
-package io.shiftleft.layers.enhancedbase
+package io.shiftleft.layers
 
 import gremlin.scala.ScalaGraph
-import io.shiftleft.codepropertygraph.generated.Languages
-import io.shiftleft.passes.{CpgPass, CpgPassRunner}
-import io.shiftleft.passes.linking.linker.Linker
-import io.shiftleft.passes.linking.memberaccesslinker.MemberAccessLinker
-import io.shiftleft.passes.namespacecreator.NamespaceCreator
 import io.shiftleft.SerializedCpg
-import io.shiftleft.semanticsloader.Semantics
+import io.shiftleft.codepropertygraph.generated.Languages
 import io.shiftleft.passes.containsedges.ContainsEdgePass
 import io.shiftleft.passes.languagespecific.fuzzyc.{MethodStubCreator, TypeDeclStubCreator}
 import io.shiftleft.passes.linking.capturinglinker.CapturingLinker
+import io.shiftleft.passes.linking.linker.Linker
+import io.shiftleft.passes.linking.memberaccesslinker.MemberAccessLinker
 import io.shiftleft.passes.methoddecorations.MethodDecoratorPass
-import io.shiftleft.passes.propagateedges.PropagateEdgePass
-import io.shiftleft.passes.reachingdef.ReachingDefPass
+import io.shiftleft.passes.namespacecreator.NamespaceCreator
 import io.shiftleft.passes.receiveredges.ReceiverEdgePass
+import io.shiftleft.passes.{CpgPass, CpgPassRunner}
 
-class EnhancedBaseCreator(graph: ScalaGraph, language: String, serializedCpg: SerializedCpg, semantics: Semantics) {
-
+class EnhancedBaseCreator(graph: ScalaGraph, language: String, serializedCpg: SerializedCpg) {
   private val runner = new CpgPassRunner(serializedCpg)
-
   private val enhancementExecList = createEnhancementExecList(language)
 
   private def createEnhancementExecList(language: String): List[CpgPass] = {
@@ -28,13 +23,11 @@ class EnhancedBaseCreator(graph: ScalaGraph, language: String, serializedCpg: Se
         List(
           new ReceiverEdgePass(graph),
           new MethodDecoratorPass(graph),
-          new PropagateEdgePass(graph, semantics),
           new CapturingLinker(graph),
           new Linker(graph),
           new MemberAccessLinker(graph),
           new ContainsEdgePass(graph),
-          new NamespaceCreator(graph),
-          new ReachingDefPass(graph),
+          new NamespaceCreator(graph)
         )
       case Languages.C =>
         List(
@@ -42,25 +35,21 @@ class EnhancedBaseCreator(graph: ScalaGraph, language: String, serializedCpg: Se
           new MethodStubCreator(graph),
           new ReceiverEdgePass(graph),
           new MethodDecoratorPass(graph),
-          new PropagateEdgePass(graph, semantics),
           new CapturingLinker(graph),
           new Linker(graph),
           new MemberAccessLinker(graph),
           new ContainsEdgePass(graph),
-          new NamespaceCreator(graph),
-          new ReachingDefPass(graph),
+          new NamespaceCreator(graph)
         )
       case _ =>
         List(
           new ReceiverEdgePass(graph),
           new MethodDecoratorPass(graph),
-          new PropagateEdgePass(graph, semantics),
           new CapturingLinker(graph),
           new Linker(graph),
           new MemberAccessLinker(graph),
           new ContainsEdgePass(graph),
-          new NamespaceCreator(graph),
-          new ReachingDefPass(graph),
+          new NamespaceCreator(graph)
         )
     }
   }
