+++ semanticcpg/src/main/scala/io/shiftleft/layers/EnhancementRunner.scala
@@ -2,13 +2,10 @@ package io.shiftleft.layers
 
 import io.shiftleft.SerializedCpg
 import io.shiftleft.codepropertygraph.Cpg
-import io.shiftleft.layers.enhancedbase.EnhancedBaseCreator
-import io.shiftleft.semanticsloader.Semantics
 
-class EnhancementRunner(semantics: Semantics) {
+class EnhancementRunner {
   def run(cpg: Cpg, serializedCpg: SerializedCpg): Unit = {
-
     val language = cpg.metaData.language.head
-    new EnhancedBaseCreator(cpg.graph, language, serializedCpg, semantics).create
+    new EnhancedBaseCreator(cpg.graph, language, serializedCpg).create
   }
 }
