+++ semanticcpg/src/test/scala/io/shiftleft/semanticcpg/Fixture.scala
@@ -4,14 +4,15 @@ import gremlin.scala._
 import io.shiftleft.SerializedCpg
 import io.shiftleft.codepropertygraph.cpgloading.{CpgLoader, CpgLoaderConfig}
 import io.shiftleft.passes.methoddecorations.MethodDecoratorPass
-import io.shiftleft.layers.EnhancementRunner
+import io.shiftleft.layers.{DataFlowRunner, EnhancementRunner}
 import io.shiftleft.semanticsloader.SemanticsLoader
 import org.apache.tinkerpop.gremlin.structure.Graph
 
 class Fixture(projectName: String) {
   val loadConfig = CpgLoaderConfig.default.copy(ignoredProtoEntries = IgnoredCpgEntities.forJava2Cpg)
   val cpg = CpgLoader.load(s"resources/cpgs/$projectName/cpg.bin.zip", loadConfig)
-  new EnhancementRunner(SemanticsLoader.emptySemantics).run(cpg, new SerializedCpg())
+  new EnhancementRunner().run(cpg, new SerializedCpg())
+  new DataFlowRunner(SemanticsLoader.emptySemantics)
   val scalaGraph: ScalaGraph = cpg.graph
 
   protected def applyMethodDecorator(graph: Graph): Unit = {
